chore(core): first wizzard iteration
